Removed legacy code in favor of ACSet progress
