+++ src/AlgebraicRelations.jl
@@ -1,8 +1,5 @@
 module AlgebraicRelations
-  include("ACSetDB.jl")
-  include("ACSetQueries.jl")
-  include("ACSetInterface.jl")
-  include("Query.jl")
-  include("SQL.jl")
+  include("DB.jl")
+  include("Queries.jl")
   include("Interface.jl")
 end
